cmd/search: Add --job-uri-prefix
And pass it around, instead of hard-coding the prefix in a number of
locations.  As part of consuming the prefix URI, extend MetadataFor to
handle the path decoding and job-URI construction as well, which
allows us to centralize some code that was previously duplicated
between renderWithContext and renderSummary.

As part of this change, I've adjusted the output format a bit.  Most
obviously by removing template.HTMLEscapeString(string(parts[3]) from
some renderWithContext output.  For example, master produced entries
like:

  build-log.txt from PR 22369 pull-ci-openshift-origin-master-e2e-aws-serial #5405 About an hour
  build-log.txt from PR openshift_installer pull-ci-openshift-installer-master-e2e-openstack #189 4 hours
  build-log.txt from build release-openshift-origin-installer-e2e-aws-upgrade #859 3 hours

With this commit, we get entries like:

  build-log from build release-openshift-origin-installer-e2e-aws-upgrade #948 7 hours
  build-log from pull pull-ci-openshift-machine-config-operator-master-e2e-aws-upgrade #104 8 hours
  build-log from pull pull-ci-openshift-hive-master-e2e #709 9 hours

The PR number is nice, but when the PR number is instead a repository
slug, it's redundant with the job name.  In fact, the trigger string
is also fairly redundant with the job name (e.g. "...from pull
pull-ci-...", but I'm leaving that alone for now because at least we
consistently detect it (vs. flip-flopping between PR number and repo
slugs).

I've also renamed --deck-url to the more generic --deck-uri.  We don't
care if these are generically locate-able or not.