cmd/search/chart: Add description and og:image <meta>
Docs in [1,2,3].  I dunno if Slack or other unfurlers execute body
JavaScript before scraping these <meta> tags, so I'm computing them
server-side.

There's a little drift between the description counts and the plotted
counts, because the description counts include all jobs which finished
in the MaxAge time range, while the plotted counts include all jobs
which started in the MaxAge time range.  Jobs that started before the
time range but finished inside it will be included in the description
counts but not in the plotted counts.

I've also noticed that point highlighting isn't working for regexps
with backslashed characters, but am just dropping a FIXME here.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/meta
[2]: https://html.spec.whatwg.org/multipage/semantics.html#meta-description
[3]: http://ogp.me/#metadata