cmd/search/chart: Only color matching failed jobs
The production CI-search deployment also indexes build logs and JUnit
files for successful jobs.  That's useful for identifying noisy
regular expressions, but the chart UI doesn't currently have a way to
distinguish between "failed and matches" and "success and matches".
Until it grows a way, just color the failures and leave the successes
gray.