Switch to indexing jobs based on a list of input prow jobs
Maintain an informer that periodically polls deck for all jobs, then
notifies an extract worker to retrieve jobs that show up. The directory
on disk is set to the start timestamp until extraction completes and
then is set to the prow job completion timestamp so duplicate jobs will
be ignored. This will not capture historical records and a future change
should monitor that.

Add a hard cap on the number of paths in the index to avoid "argument
list too long" errors. Move some logic out of the index and into the
top level argument construction so that we can fall back to searching
everything.

Start using the .Done() method on queues - that was an oversight and
would lead to unbounded memory growth in some cases.