Use the new job index created in the bucket to backfill old jobs
We can use newer prow jobs for incremental reads and backfill from
the index on startup. The index is maintained by cloud functions in
github.com/openshift/ci-search-functions and has a time sharded
key index. We scan through that set and add them to the work queue.

In addition, we change to index junit.failures from successful jobs
as well, and force build-logs.txt to be no more than 20MB.