prow/types: Make completionTime omitempty
To match [1], fixing a regression introduced by 06c45ec87f (Switch to
indexing jobs based on a list of input prow jobs, 2020-03-28, #44).
This fixes the durations for the pending jobs.  The current JavaScript
in cmd/search/httpchart.go has:

  if (job.status.completionTime === undefined) {
    job.finished = now;
  }  else {
    job.finished = isoParse(job.status.completionTime);
  }

so for pending jobs to get the 'now' default, they must have
completionTime unset.

[1]: https://github.com/kubernetes/test-infra/blob/e35a6bca1f57be69cbe0a24b0b6c076407b2e01f/prow/apis/prowjobs/v1/types.go#L624-L625