Go back to using the path index
Without the path index we are at risk of page cache eviction and
subsequent thrashing when we have to scan the whole directory. To
get back to using the path index we have to support scanning the
input in chunks based on the max path size for fork/exec (lower on
Mac). At the same time, remove a few other common allocators and
streamline the search path.