Ensure the pathindex filters by job name
This massively speeds up deep queries on job names and is a prerequisite
for complex job searches. At the same time revert the job searches in
the HTTP methods to match the implementation.

A side effect of this change is that bug searchs will search on the
pretty title, which is desirable in some cases (show all bugs with this
name).