Cap maximum bytes returned and abort search if a write fails
When clients disconnected from the server rg would continue running
until it had found all results. Change the grep function to return
an error and use that to terminate search (and thus rg) early if
such a disconnection happens.

Support a cap on the maximum bytes returned by the search determined
by the maxBytes argument and default it to 20Mb. If this limit is
hit the current result is returned and then an error is printed,
which also prevents display of the search summary (future improvement).