Add a group by job filter that shows the frequency of the match
This will allow us to estimate impact of a specific search result
more accurately and track how it may change over time. Requires that
we fill out the jobs loaded from the index and then pass that through
into the cache so we can check the total list of jobs quickly.

Try to share as much render code as possible with the current
search view.