prow: The jobs informer stops reporting prow jobs after 24 hours
The jobs informer uses the GCS index to get an initial set for the
full range, but relies on periodic scrapes of deck to get the
current set of jobs. When jobs expired from deck they were no longer
correctly reporting stats for the new jobs based searches which led
to incorrect estimates.

The list watcher for the informer now remembers the last set of
jobs it returned and merges the updated list with that set, expiring
jobs older than the MaxAge. A new metric is added which tracks the
high water mark for jobs in the cache to better diagnose this in
the future.