bugzilla: collect version information too
This changes search.ci to collect the Version information from Bugzilla
too.  The reason is new bugs in Sippy end up showing as associated, not
linked, to a release.  If Target release is blank (e.g. for a new bug),
I want Sippy to consider the version too. That way new bugs show up
linked.