metricsdb/httpgraph: adjust decimal display
When dealing with latency metrics, values are using second as a unit and
we want to know their value up to the milisecond. So to make it easier
to understand variation on the graph, we will display value up to 3
decimals.

Since the order of magnitude differ depending on the scenario, value < 1
will have 3 decimals displayed, < 10 2 decimals displayed and < 100 1
decimal displayed.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>